+++ go/cmd/vtctld/vtctld.go
@@ -2,36 +2,13 @@ package main
 
 import (
 	"flag"
-	"fmt"
-	"net/http"
-	"path"
-	"strings"
 
-	"golang.org/x/net/context"
-
-	log "github.com/golang/glog"
-	"github.com/youtube/vitess/go/acl"
 	"github.com/youtube/vitess/go/vt/servenv"
 	"github.com/youtube/vitess/go/vt/topo"
-	"github.com/youtube/vitess/go/vt/wrangler"
-
-	topodatapb "github.com/youtube/vitess/go/vt/proto/topodata"
-)
-
-var (
-	webDir = flag.String("web_dir", "", "directory from which to serve vtctld web interface resources")
-
-	appPrefix = "/app/"
 )
 
 func init() {
 	servenv.RegisterDefaultFlags()
-}
-
-func httpErrorf(w http.ResponseWriter, r *http.Request, format string, args ...interface{}) {
-	errMsg := fmt.Sprintf(format, args...)
-	log.Errorf("HTTP error on %v: %v, request: %#v", r.URL.Path, errMsg, r)
-	http.Error(w, errMsg, http.StatusInternalServerError)
 }
 
 // used at runtime by plug-ins
@@ -47,104 +24,8 @@ func main() {
 	ts = topo.GetServer()
 	defer topo.CloseServers()
 
-	actionRepo := NewActionRepository(ts)
-
-	// keyspace actions
-	actionRepo.RegisterKeyspaceAction("ValidateKeyspace",
-		func(ctx context.Context, wr *wrangler.Wrangler, keyspace string, r *http.Request) (string, error) {
-			return "", wr.ValidateKeyspace(ctx, keyspace, false)
-		})
-
-	actionRepo.RegisterKeyspaceAction("ValidateSchemaKeyspace",
-		func(ctx context.Context, wr *wrangler.Wrangler, keyspace string, r *http.Request) (string, error) {
-			return "", wr.ValidateSchemaKeyspace(ctx, keyspace, nil, false)
-		})
-
-	actionRepo.RegisterKeyspaceAction("ValidateVersionKeyspace",
-		func(ctx context.Context, wr *wrangler.Wrangler, keyspace string, r *http.Request) (string, error) {
-			return "", wr.ValidateVersionKeyspace(ctx, keyspace)
-		})
-
-	actionRepo.RegisterKeyspaceAction("ValidatePermissionsKeyspace",
-		func(ctx context.Context, wr *wrangler.Wrangler, keyspace string, r *http.Request) (string, error) {
-			return "", wr.ValidatePermissionsKeyspace(ctx, keyspace)
-		})
-
-	// shard actions
-	actionRepo.RegisterShardAction("ValidateShard",
-		func(ctx context.Context, wr *wrangler.Wrangler, keyspace, shard string, r *http.Request) (string, error) {
-			return "", wr.ValidateShard(ctx, keyspace, shard, false)
-		})
-
-	actionRepo.RegisterShardAction("ValidateSchemaShard",
-		func(ctx context.Context, wr *wrangler.Wrangler, keyspace, shard string, r *http.Request) (string, error) {
-			return "", wr.ValidateSchemaShard(ctx, keyspace, shard, nil, false)
-		})
-
-	actionRepo.RegisterShardAction("ValidateVersionShard",
-		func(ctx context.Context, wr *wrangler.Wrangler, keyspace, shard string, r *http.Request) (string, error) {
-			return "", wr.ValidateVersionShard(ctx, keyspace, shard)
-		})
-
-	actionRepo.RegisterShardAction("ValidatePermissionsShard",
-		func(ctx context.Context, wr *wrangler.Wrangler, keyspace, shard string, r *http.Request) (string, error) {
-			return "", wr.ValidatePermissionsShard(ctx, keyspace, shard)
-		})
-
-	// tablet actions
-	actionRepo.RegisterTabletAction("Ping", "",
-		func(ctx context.Context, wr *wrangler.Wrangler, tabletAlias *topodatapb.TabletAlias, r *http.Request) (string, error) {
-			ti, err := wr.TopoServer().GetTablet(ctx, tabletAlias)
-			if err != nil {
-				return "", err
-			}
-			return "", wr.TabletManagerClient().Ping(ctx, ti)
-		})
-
-	actionRepo.RegisterTabletAction("RefreshState", acl.ADMIN,
-		func(ctx context.Context, wr *wrangler.Wrangler, tabletAlias *topodatapb.TabletAlias, r *http.Request) (string, error) {
-			ti, err := wr.TopoServer().GetTablet(ctx, tabletAlias)
-			if err != nil {
-				return "", err
-			}
-			return "", wr.TabletManagerClient().RefreshState(ctx, ti)
-		})
-
-	actionRepo.RegisterTabletAction("DeleteTablet", acl.ADMIN,
-		func(ctx context.Context, wr *wrangler.Wrangler, tabletAlias *topodatapb.TabletAlias, r *http.Request) (string, error) {
-			return "", wr.DeleteTablet(ctx, tabletAlias, false, false)
-		})
-
-	actionRepo.RegisterTabletAction("ReloadSchema", acl.ADMIN,
-		func(ctx context.Context, wr *wrangler.Wrangler, tabletAlias *topodatapb.TabletAlias, r *http.Request) (string, error) {
-			return "", wr.ReloadSchema(ctx, tabletAlias)
-		})
-
-	// Anything unrecognized gets redirected to the main app page.
-	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		http.Redirect(w, r, appPrefix, http.StatusFound)
-	})
-
-	// Serve the static files for the vtctld web app.
-	http.HandleFunc(appPrefix, func(w http.ResponseWriter, r *http.Request) {
-		// Strip the prefix.
-		parts := strings.SplitN(r.URL.Path, "/", 3)
-		if len(parts) != 3 {
-			http.NotFound(w, r)
-			return
-		}
-		rest := parts[2]
-		if rest == "" {
-			rest = "index.html"
-		}
-		http.ServeFile(w, r, path.Join(*webDir, rest))
-	})
-
-	// Serve the REST API for the vtctld web app.
-	initAPI(context.Background(), ts, actionRepo)
-
-	// Init redirects for explorers
-	initExplorer()
+	// Init the vtctld core
+	InitVtctld(ts)
 
 	// Start schema manager service.
 	initSchema()
